Fixed some issues with encoding data length for ver.10+

+++ pkg/qr-encode/encode_test.go
@@ -66,13 +66,13 @@ func Test_fillBuffer(t *testing.T) {
 	require.Equal(t, versionSize[e.level][version], buff.Len()*8)
 
 	b := buff.Bytes()
-	header := int(b[0] >> 4)
+	header := b[0] >> 4
 	require.Equal(t, headerNibble, header)
 
-	actualLen := int(byte(int(b[0])&Nible) | (b[1] >> 4))
+	actualLen := int(byte(b[0]&Nible) | byte(b[1]>>4))
 	require.Equal(t, len(data), actualLen)
 
-	lastByte := b[10] | byte(int(b[9])&Nible)
+	lastByte := b[10] | byte(b[9]&Nible)
 	require.Equal(t, data[len(data)-1], lastByte)
 
 }
